+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud/compare/v3.0.1...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+Fixes `pyaud format-str`
 
 [3.0.1](https://github.com/jshwi/pyaud/releases/tag/v3.0.1) - 2021-07-24
 ------------------------------------------------------------------------