+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud/compare/v3.0.2...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+Reduces indexing time
 
 [3.0.2](https://github.com/jshwi/pyaud/releases/tag/v3.0.2) - 2021-07-25
 ------------------------------------------------------------------------