+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud/compare/v3.0.3...HEAD)
 ------------------------------------------------------------------------
+### Added
+Adds all git commands to `pyaud.git`
 
 [3.0.3](https://github.com/jshwi/pyaud/releases/tag/v3.0.3) - 2021-07-26
 ------------------------------------------------------------------------